+++ src/views/monitor/job/job-list.vue
@@ -267,6 +267,7 @@
   } from '@/api/monitor/job'
   import Pagination from '@/components/Pagination/index' // secondary package based on el-pagination
   import jobAddUpdateView from '@/views/monitor/job/job-add-update-view'
+  import { MessageBox } from '_element-ui@2.13.0@element-ui'
 
   export default {
     name: 'userList',
@@ -527,6 +528,12 @@
         }
       },
       handleDeleteJobs() {
+        if(this.multipleSelection.length==0){
+          MessageBox.confirm('请选择至少一条数据', '批量删除数据', {
+            type: 'warning'
+          })
+          return
+        }
         deleteJobs({ ids: (this.multipleSelection) + '' }).then(data => {
           if (data.code == '20000') {
             this.$notify({
